requirements.txt
asgiref==3.4.1
Django==3.2.5
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-js-asset==1.2.2
django-taggit==1.5.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
Pillow==8.3.0
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
